HomeWork1 Done
