+++ src/main/java/LevelTwoJava/HomeWorkOne/App.java
@@ -1,4 +1,4 @@
-package LevelTwoJava.HomeWork1;
+package LevelTwoJava.HomeWorkOne;
 
 public class App {
     public static void main(String[] args) {
