+++ src/main/java/LevelTwoJava/HomeWorkOne/Cat.java
@@ -1,4 +1,4 @@
-package LevelTwoJava.HomeWork1;
+package LevelTwoJava.HomeWorkOne;
 
 public class Cat implements Players {
